+++ src/core/state.hpp
@@ -55,8 +55,8 @@ class State : public godot::RefCounted
 		godot::PackedInt32Array bit_index(int _piece);
 		int get_turn();
 		void set_turn(int _turn);
-		int64_t get_castle();
-		void set_castle(int64_t _castle);
+		int get_castle();
+		void set_castle(int _castle);
 		int get_en_passant();
 		void set_en_passant(int _en_passant);
 		int get_step_to_draw();
@@ -75,18 +75,6 @@ class State : public godot::RefCounted
 	private:
 		int pieces[128] = {0};
 		int64_t bit[128] = {0};
-		//'*'表示所有棋子
-		//'!'旋转90度后的棋盘
-		//'('逆时针旋转45度的棋盘
-		//')'顺时针旋转45度的棋盘
-		//int turn = 0;
-		//int castle = 0xF;
-		//int en_passant = -1;
-		//int step_to_draw = 0;
-		//int round = 1;
-		//int king_passant = -1;
-		//int pawn_dir = 0x61;	//兵的前进方向，前4位为白方，后4位为黑方
-		//int64_t zobrist = 0;
 };
 
 #endif
